+++ src/main/java/org/example/backend/models/items/OtherItem.java
@@ -2,6 +2,7 @@ package org.example.backend.models.items;
 
 import org.example.backend.models.Item;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class OtherItem extends Item {
@@ -9,7 +10,7 @@ public class OtherItem extends Item {
     protected String origin;   // Xuất xứ
     protected double weight;   // Trọng lượng (kg)
 
-    public OtherItem(int id, LocalDateTime createdAt, String type, String itemName,String category, String origin, double weight, String description, double startingPrice) {
+    public OtherItem(int id, LocalDateTime createdAt, String type, String itemName,String category, String origin, double weight, String description, BigDecimal startingPrice) {
         super(id, createdAt, type, itemName, description, startingPrice);
         this.category = category;
         this.origin = origin;
